add reach slider
