+++ src/components/canvas/Editor/ColorPicker.js
@@ -1,0 +1,33 @@
+import { ChromePicker } from 'react-color'
+
+const ColorPicker = ({ property, value, material, setMaterialsEditor }) => {
+  const onChange = (val) => {
+    if (val.hex) {
+      setMaterialsEditor((m) => ({
+        ...m,
+        [material]: {
+          ...m[material],
+          color: {
+            type: 'color',
+            value: val.hex,
+          },
+        },
+      }))
+    }
+  }
+  return (
+    <>
+      <span className='block mb-2 font-medium text-gray-800 capitalize'>
+        Color
+      </span>
+      <ChromePicker
+        key={property}
+        disableAlpha={true}
+        color={value}
+        onChange={onChange}
+      />
+    </>
+  )
+}
+
+export default ColorPicker
