+++ src/components/canvas/Editor/EditTools.js
@@ -1,0 +1,94 @@
+import Button from '@/components/Button'
+import classNames from '@/helpers/classNames'
+import { useState } from 'react'
+
+import { saveAs } from 'file-saver'
+import { GLTFExporter } from 'three-stdlib'
+import { useRouter } from 'next/router'
+import RangeSlider from './RangeSlider'
+import ColorPicker from './ColorPicker'
+import { useUpdateScene } from './useUpdateScene'
+
+const EditTools = ({ materialsEditor, scene, setMaterialsEditor }) => {
+  const [activeTab, setActiveTab] = useState()
+  const exporter = new GLTFExporter()
+  const router = useRouter()
+  useUpdateScene({ materialsEditor, scene })
+
+  const save = () => {
+    exporter.parse(scene, function (gltf) {
+      var blob = new Blob([JSON.stringify(gltf)], {
+        type: 'application/gltf; charset=us-ascii',
+      })
+      saveAs(blob, 'model.gltf')
+    })
+  }
+
+  return (
+    <main className='p-4'>
+      <button onClick={() => router.back()}>
+        <svg
+          xmlns='http://www.w3.org/2000/svg'
+          className='h-10 w-10'
+          fill='none'
+          viewBox='0 0 24 24'
+          stroke='currentColor'
+        >
+          <path
+            strokeLinecap='round'
+            strokeLinejoin='round'
+            strokeWidth={2}
+            d='M11 17l-5-5m0 0l5-5m-5 5h12'
+          />
+        </svg>
+      </button>
+      <div className='flex mt-10 background-white'>
+        <nav className='mr-5 space-y-1'>
+          {Object.keys(materialsEditor).map((item) => (
+            <button
+              key={item.name}
+              onClick={() => setActiveTab(item)}
+              className={classNames(
+                activeTab === item
+                  ? 'bg-gray-100 text-gray-900'
+                  : 'text-gray-600 hover:bg-gray-50 hover:text-gray-900',
+                'flex items-center px-3 py-2 text-sm font-medium rounded-md'
+              )}
+            >
+              <span className='capitalize truncate'>{item}</span>
+            </button>
+          ))}
+        </nav>
+        {Object.keys(materialsEditor).map((material) => (
+          <div
+            key={material}
+            className={`${activeTab === material ? '' : 'hidden'}`}
+          >
+            {Object.keys(materialsEditor[material]).map((property) => {
+              const val = materialsEditor[material][property]
+              const props = {
+                material,
+                property,
+                setMaterialsEditor,
+                value: val.value,
+              }
+              if (val.type === 'range') {
+                return <RangeSlider {...props} />
+              }
+              if (val.type === 'color') {
+                return <ColorPicker {...props} />
+              }
+
+              return null
+            })}
+          </div>
+        ))}
+      </div>
+      <Button className='fixed bottom-5 left-5' onClick={save}>
+        Download Model
+      </Button>
+    </main>
+  )
+}
+
+export default EditTools
